V0.2.0 rc.

+++ lib/ADC/ADC.h
@@ -1,10 +1,16 @@
 /**
- * ADC.c
- * Devices: LM4F120; TM4C123
- * Description: Low level drivers for ADC initialization.
- * Authors: Matthew Yu.
- * Last Modified: 09/15/21
+ * @file ADC.h
+ * @author Matthew Yu (matthewjkyu@gmail.com)
+ * @brief ADC peripheral driver.
+ * @version 0.1
+ * @date 2021-09-22
+ * @copyright Copyright (c) 2021
+ * @note
+ * Unsupported Features. This driver does not support DMA control. This driver
+ * does not support different trigger sources or interrupts. The driver does not
+ * support configurable sample sequencer priorities.
  */
+
 #pragma once
 
 /** General imports. */
@@ -13,6 +19,9 @@
 /** Device specific imports. */
 #include <lib/GPIO/GPIO.h>
 
+
+/** @brief ADCPin is an enumeration that specifies one of the available ADC pins
+ *         on the TM4C. */
 enum ADCPin {
     AIN0, // PE3
     AIN1, // PE2
@@ -28,11 +37,14 @@ enum ADCPin {
     AIN11 // PB5
 };
 
+/** @brief ADCModule is an enumeration specifying of the two ADC modules. */
 enum ADCModule {
     ADC_MODULE_0,
     ADC_MODULE_1
 };
 
+/** @brief ADCSequencer is an enumeration specifying of the four ADC sequencers
+ *         that exist for each ADCModule. */
 enum ADCSequencer {
     ADC_SS_0,
     ADC_SS_1,
@@ -40,6 +52,9 @@ enum ADCSequencer {
     ADC_SS_3
 };
 
+/** @brief ADCSequencePosition is an enumeration specifying of the eight
+ *         slots in an ADC sequencer. Some sequencers may not have all eight
+ *         positions available. */
 enum ADCSequencePosition {
     ADC_SEQPOS_0,
     ADC_SEQPOS_1,
@@ -51,6 +66,9 @@ enum ADCSequencePosition {
     ADC_SEQPOS_7
 };
 
+/** @brief ADCAveraging is an enumeration specifying how many samples are taken
+ *         in succession before returning an averaged result. May extend
+ *         execution time when sampling. */
 enum ADCAveraging {
     ADC_AVG_NONE,
     ADC_AVG_1,
@@ -62,6 +80,8 @@ enum ADCAveraging {
     ADC_AVG_64
 };
 
+/** @brief ADCPhase is an enumeration specifying the phase offset that a given
+ *         ADCModule is sampling at. */
 enum ADCPhase {
     ADC_PHASE_0,
     ADC_PHASE_22_5,
@@ -81,10 +101,12 @@ enum ADCPhase {
     ADC_PHASE_337_5
 };
 
-/** Configuration of a single DAC object. */
+/** @brief ADCConfig_t is a user defined struct that specifies an ADC pin
+ *         configuration. */
 typedef struct ADCConfig {
     /**
-     * The GPIO that can be selected as ADC pins.
+     * @brief The GPIO pin that can be selected as ADC pins.
+     * 
      * Default is AIN0, which corresponds to PE3.
      */
     enum ADCPin pin;
@@ -93,69 +115,74 @@ typedef struct ADCConfig {
     /** ------------- Optional Fields. ------------- */
 
     /**
-     * The ADC module associated with the ADC Pin. There are two ADC modules
-     * in the TM4C123GH6PM and both can work simultaneously.
+     * @brief The ADC module associated with the ADC Pin. There are two ADC
+     * modules in the TM4C123GH6PM and both can operate simultaneously.
      *
      * Default ADC_MODULE_0.
      */
     enum ADCModule module;
 
-
     /**
-     * The ADC sequencer associated with the ADC Pin. There are four sequencers
-     * PER ADC module. They do not work simultaneously relative to other sequencers
-     * in the module. Multiple enabled sequencers are given priority in reverse
-     * order (e.g. SS0 goes after SS3).
+     * @brief The ADC sequencer associated with the ADC Pin. There are four sequencers
+     *        PER ADC module. They do not work simultaneously relative to other
+     *        sequencers in the module. Multiple enabled sequencers are given
+     *        priority in reverse order (e.g. SS0 goes after SS3).
      *
      * Default ADC_SS_0.
      */
     enum ADCSequencer sequencer;
 
     /**
-     * The position in the sample sequencer that the ADCPin is sampled. Sequencers
-     * may have the ability to sample multiple pins or the same pins in one go -
-     * the position dictates where in the order of iteration the sample is captured.
+     * @brief The position in the sample sequencer that the ADCPin is sampled.
+     *        Sequencers may have the ability to sample multiple pins or the
+     *        same pins in one go - the position dictates where in the order of
+     *        iteration the sample is captured.
      *
      * Default ADC_SEQPOS_0.
+     * 
+     * @note Different sequencers have different number of positions.
+     *       ADC_SS_0 has all 8 positions, ADC_SS_1 and ADC_SS_2 have only 4
+     *       positions, and ADC_SS_3 has a single position.
      */
     enum ADCSequencePosition position;
 
     /**
-     * Whether the sample to configure is the last in a series of samples in the
-     * sequencer to be read.
+     * @brief Whether the sample to configure is the last in a series of samples
+     * in the sequencer to be read.
      *
-     * Default False (It IS the end sample).
+     * Default is false (It IS the end sample).
      */
     bool isNotEndSample;
 
     /**
-     * The hardware sample averaging. This allows the ADC to take multiple samples
-     * at once before merging them into a single entry that goes into the internal
-     * FIFO that the user can pull. Hardware sample averaging smooths out noise
-     * over consecutive samples and improves the accuracy of the result (see:
-     * Central Limit Theorem) at the expense of power and more processing time.
+     * @brief The hardware sample averaging. This allows the ADC to take
+     *        multiple samples at once before merging them into a single entry
+     *        that goes into the internal FIFO that the user can pull. Hardware
+     *        sample averaging smooths out noise over consecutive samples and
+     *        improves the accuracy of the result (see: Central Limit Theorem)
+     *        at the expense of power and more processing time. 
      *
      * Default ADC_AVG_NONE.
      */
     enum ADCAveraging oversampling;
 
     /**
-     * Whether the ADC samples are dithered or not. Dithering may reduce random
-     * noise in the ADC sampling, especially when combined with hardware sample
-     * averaging.
+     * @brief Whether the ADC samples are dithered or not. Dithering may reduce
+     *        random noise in the ADC sampling, especially when combined with
+     *        hardware sample averaging.
      *
      * Default false.
      */
     bool isDithered;
 
     /**
-     * ADC Sample Phase Control. This causes the samples to lag by a set phase.
-     * This can be used to effectively double the ADC sample rate. See P.805
-     * of the datasheet.
+     * @brief ADC Sample Phase Control. This causes the samples to lag by a set
+     *        phase. This can be used to effectively double the ADC sample rate.
+     *        See P.805 of the datasheet.
+     * 
      * Default 0.0 degrees.
      */
     enum ADCPhase phase;
-
 
     /** ------------- Unimplemented Fields. ------------- */
 
@@ -173,28 +200,28 @@ typedef struct ADCConfig {
     // void (*ADCHandler)(void);
 } ADCConfig_t;
 
+/** @brief ADC_t is a struct containing user relevant data of an ADC. */
 typedef struct ADC {
     /**
-     * The GPIO that can be selected as ADC pins.
+     * @brief The GPIO that can be selected as ADC pins.
+     * 
      * Default is AIN0, which corresponds to PE3.
-     * The user may modify this to a unspecified pin value - this will cause
-     * undefined behavior.
      */
     enum ADCPin pin;
 
-    /** The ADCModule associated with the ADC Pin. */
+    /** @brief The ADCModule associated with the ADC Pin. */
     enum ADCModule module;
 
-    /** The ADCSequencer associated with the ADC Pin. */
+    /** @brief The ADCSequencer associated with the ADC Pin. */
     enum ADCSequencer sequencer;
 
-    /** The ADCSequencePosition associated with the ADC Pin. */
+    /** @brief The ADCSequencePosition associated with the ADC Pin. */
     enum ADCSequencePosition position;
 } ADC_t;
 
 
 /**
- * ADCInit initializes an ADC module given an ADCConfig_t configuration.
+ * @brief ADCInit initializes an ADC module given an ADCConfig_t configuration.
  * 
  * @param config The configuration of the ADC module.
  * @return An ADC_t struct instance used for sampling.
@@ -202,8 +229,8 @@ typedef struct ADC {
 ADC_t ADCInit(ADCConfig_t config);
 
 /**
- * ADCIsEmpty checks whether the sequencer at the given ADC module has no
- * samples available.
+ * @brief ADCIsEmpty checks whether the sequencer at the given ADC module has no
+ *        samples available.
  * 
  * @param module ADC module to check.
  * @param sequencer ADC module's sequencer to check.
@@ -212,8 +239,8 @@ ADC_t ADCInit(ADCConfig_t config);
 bool ADCIsEmpty(enum ADCModule module, enum ADCSequencer sequencer);
 
 /**
- * ADCIsFull checks whether the sequencer at the given ADC module cannot capture
- * any more samples.
+ * @brief ADCIsFull checks whether the sequencer at the given ADC module cannot
+ *        capture any more samples.
  * 
  * @param module ADC module to check.
  * @param sequencer ADC module's sequencer to check.
@@ -222,22 +249,24 @@ bool ADCIsEmpty(enum ADCModule module, enum ADCSequencer sequencer);
 bool ADCIsFull(enum ADCModule module, enum ADCSequencer sequencer);
 
 /**
- * ADCSampleSingle samples a single ADC pin. This may be slower if the sequencer
- * is configured to capture other samples, as this method cycles through the
- * internal sample FIFO to capture the wanted value and throw out the others.
+ * @brief ADCSampleSingle samples a single ADC pin. This may be slower if the
+ *        sequencer is configured to capture other samples, as this method
+ *        cycles through the internal sample FIFO to capture the wanted value
+ *        and throw out the others.
  *
  * @param adc The ADC object containing the pin value that should be sampled.
- * @return A single uint32_t representing the pin ADC value sampled.
+ * @return A single uint16_t representing the pin ADC value sampled.
  */
-uint32_t ADCSampleSingle(ADC_t adc);
+uint16_t ADCSampleSingle(ADC_t adc);
 
 /**
- * ADCSampleSequencer samples a single sequencer and returns all the values in the
- * FIFO accumulated, up to 8 values. This may be faster than ADCSampleSingle on a
- * per sample basis. Guarantees that the internal FIFO is empty beforehand.
+ * @brief ADCSampleSequencer samples a single sequencer and returns all the
+ *        values in the FIFO accumulated, up to 8 values. This may be faster
+ *        than ADCSampleSingle on a  per sample basis. Guarantees that the
+ *        internal FIFO is empty beforehand.
  *
  * @param module The ADCModule to execute.
  * @param sequencer The ADCSequencer to sequence.
  * @param arr A reference to an array to fill with values.
  */
-void ADCSampleSequencer(enum ADCModule module, enum ADCSequencer sequencer, uint32_t arr[8]);
+void ADCSampleSequencer(enum ADCModule module, enum ADCSequencer sequencer, uint16_t arr[8]);
